//TEMP: deep linking implementation.
