+++ app/store/reducers/appState/navigation.js
@@ -1,5 +1,8 @@
 import AppNavigator from '../../../routes/AppNavigator';
 
 export default function reducer(state, action) {
+  if(action.routeName === 'Home' && !state.isAuthenticated) {
+    return AppNavigator.router.getStateForAction({}, state);
+  }
   return AppNavigator.router.getStateForAction(action, state);
 }
